+++ materialize-s3-parquet/CHANGELOG.md
@@ -1,4 +1,9 @@
 # materialize-s3-parquet
 
+## v2, 2022-09-22
+
+- Moved endpoint config to advanced section.
+
 ## v1, 2022-07-27
+
 - Beginning of changelog.